code/pyproject.toml
[project]
name = "code"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11, <3.12"
dependencies = [
    "einops>=0.8.1",
    "matplotlib>=3.10.8",
    "numpy>=2.4.0",
    "torchdiffeq>=0.2.5",
    "tqdm>=4.67.1",
    "wandb>=0.23.1",
]
